// packages/sdk-outpost/src/contracts/ethereum/generated/ReserveManager.ts
export declare namespace ReserveManagerLib {
  export type ReserveCreateArgsStruct = {
    tokenCode: BigNumberish;
    reserveCode: BigNumberish;
    externalTokenAmount: BigNumberish;
    requestedWireAmount: BigNumberish;
    connectorWeightBps: BigNumberish;
    name: string;
    description: string;
    isPrivate: boolean;
    creatorPubKey: BytesLike;
  };

  export type ReserveCreateArgsStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    number,
    string,
    string,
    boolean,
    string
  ] & {
    tokenCode: BigNumber;
    reserveCode: BigNumber;
    externalTokenAmount: BigNumber;
    requestedWireAmount: BigNumber;
    connectorWeightBps: number;
    name: string;
    description: string;
    isPrivate: boolean;
    creatorPubKey: string;
  };

  export type PermitSigStruct = {
    deadline: BigNumberish;
    v: BigNumberish;
    r: BytesLike;
    s: BytesLike;
  };

  export type PermitSigStructOutput = [BigNumber, number, string, string] & {
    deadline: BigNumber;
    v: number;
    r: string;
    s: string;
  };

  export type SwapArgsStruct = {
    sourceTokenCode: BigNumberish;
    sourceReserveCode: BigNumberish;
    sourceAmount: BigNumberish;
    targetChainCode: BigNumberish;
    targetTokenCode: BigNumberish;
    targetReserveCode: BigNumberish;
    targetRecipient: BytesLike;
    targetAmount: BigNumberish;
    targetToleranceBps: BigNumberish;
  };

  export type SwapArgsStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    string,
    BigNumber,
    number
  ] & {
    sourceTokenCode: BigNumber;
    sourceReserveCode: BigNumber;
    sourceAmount: BigNumber;
    targetChainCode: BigNumber;
    targetTokenCode: BigNumber;
    targetReserveCode: BigNumber;
    targetRecipient: string;
    targetAmount: BigNumber;
    targetToleranceBps: number;
  };
}